LA-63: Create a new controller for Quick Quote
