+++ src/main/java/com/selina/lending/api/controller/QuickQuoteController.java
@@ -1,0 +1,45 @@
+/*
+ * Copyright 2022 Selina Finance
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+package com.selina.lending.api.controller;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.RestController;
+
+import com.selina.lending.internal.dto.ApplicationResponse;
+import com.selina.lending.internal.dto.QuickQuoteApplicationRequest;
+
+import lombok.extern.slf4j.Slf4j;
+
+@RestController
+@Slf4j
+public class QuickQuoteController implements QuickQuoteOperations {
+
+    @Override
+    public ResponseEntity<ApplicationResponse> createQuickQuoteApplication(
+            QuickQuoteApplicationRequest quickQuoteApplicationRequest) {
+        log.info("Create Quick Quote application with [externalApplicationId={}]", quickQuoteApplicationRequest.getExternalApplicationId());
+        return ResponseEntity.ok().build();
+    }
+
+    @Override
+    public ResponseEntity<ApplicationResponse> updateQuickQuoteApplication(String externalApplicationId,
+            QuickQuoteApplicationRequest quickQuoteApplicationRequest) {
+        log.info("Update Quick Quote application with [externalApplicationId={}]", quickQuoteApplicationRequest.getExternalApplicationId());
+        return ResponseEntity.ok().build();
+    }
+}
