+++ src/test/java/com/selina/lending/api/controller/QuickQuoteControllerTest.java
@@ -1,0 +1,67 @@
+/*
+ * Copyright 2022 Selina Finance
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+package com.selina.lending.api.controller;
+
+import static org.junit.jupiter.api.Assertions.*;
+import static org.mockito.Mockito.when;
+
+import java.util.UUID;
+
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.junit.jupiter.MockitoExtension;
+
+import com.selina.lending.internal.dto.QuickQuoteApplicationRequest;
+
+@ExtendWith(MockitoExtension.class)
+class QuickQuoteControllerTest {
+
+    @InjectMocks
+    private QuickQuoteController quickQuoteController;
+
+    @Mock
+    private QuickQuoteApplicationRequest quickQuoteApplicationRequest;
+
+    @Test
+    void createQuickQuoteApplication() {
+        //Given
+        var id = UUID.randomUUID().toString();
+        when(quickQuoteApplicationRequest.getExternalApplicationId()).thenReturn(id);
+
+        //When
+        var response = quickQuoteController.createQuickQuoteApplication(quickQuoteApplicationRequest);
+
+        //Then
+        assertNotNull(response);
+    }
+
+    @Test
+    void updateQuickQuoteApplication() {
+        //Given
+        var id = UUID.randomUUID().toString();
+        when(quickQuoteApplicationRequest.getExternalApplicationId()).thenReturn(id);
+
+        //When
+        var response = quickQuoteController.updateQuickQuoteApplication(id, quickQuoteApplicationRequest);
+
+        //Then
+        assertNotNull(response);
+    }
+}
